+++ dscribe/libmbtr/mbtr.pxd
@@ -11,8 +11,8 @@ cdef extern from "mbtr.h":
   cdef cppclass MBTR:
         MBTR(vector[vector[float]], vector[int], map[int,int], int, vector[vector[int]], bool) except +
         map[string,vector[double]] getK1(string, string, map[string, float], float, float, float, float) except +
-        map[string,vector[double]] getK2(vector[vector[float]], vector[vector[int]], string, string, map[string, float], float, float, float, float) except +
-        map[string,vector[double]] getK3(vector[vector[float]], vector[vector[int]], string, string, map[string, float], float, float, float, float) except +
+        map[string,vector[double]] getK2(vector[vector[double]], vector[vector[int]], string, string, map[string, float], float, float, float, float) except +
+        map[string,vector[double]] getK3(vector[vector[double]], vector[vector[int]], string, string, map[string, float], float, float, float, float) except +
         pair[map[string,vector[float]], map[string,vector[float]]] getK1GeomsAndWeightsCython(string, string, map[string, float]) except +
         pair[map[string,vector[float]], map[string,vector[float]]] getK2GeomsAndWeightsCython(vector[vector[float]], vector[vector[int]], string, string, map[string, float]) except +
         pair[map[string,vector[float]], map[string,vector[float]]] getK3GeomsAndWeightsCython(vector[vector[float]], vector[vector[int]], string, string, map[string, float]) except +
